Cleanup: Revert Fairy test strings and resolve compile warnings

+++ include/pokemon.h
@@ -243,7 +243,7 @@ struct BattleMove
     u8  type;
     u8  accuracy;
     u8  pp;
-    u8  target;
+    u16 target;
     s8  priority;
     u8  category;
